Made clearer how MVVM will be implemented
MVVM (Model, View, ViewModel) is a design pattern that WPF is designed for. It separates out the UI (View), intermediate code to run and update the view (ViewModel) and "backend" computations that the ViewModel initiates and receives data from (Model). This decoupling allows for more flexibility and parallel development of each part of the overall application